Patch from Vernon Sauder. The off_t type is not a consistent size; it depends on the kernel options (something about large file support). Therefore, the format string for printing an address is not always the same.

+++ libbb/dump.c
@@ -650,7 +650,7 @@ static void display(void)
 		for (pr = endfu->nextpr; pr; pr = pr->nextpr) {
 			switch (pr->flags) {
 			case F_ADDRESS:
-				(void) printf(pr->fmt, eaddress);
+				(void) printf(pr->fmt, (unsigned int) eaddress);
 				break;
 			case F_TEXT:
 				(void) printf(pr->fmt);
